Building An Interactive Desktop Dictionary App With Python
This was my first Python project, one which I built as a part of Ardit Sulce’s The Python Mega Course on Udemy (a very useful course if you want to gain a holistic view of Python’s applications other than Data Science). The implementation is more or less same as how Ardit teaches in the course (at least it was two years ago), with some subtle enhancements.